refactor: server client socket example
